+++ includes/requirements.php
@@ -62,7 +62,7 @@ function wpsp_requirements_check( $force_check = false ) {
 
 
 function wpsp_notice_aihrus() {
-	$help_url  = esc_url( 'https://aihrus.zendesk.com/entries/35689458' );
+	$help_url  = esc_url( 'https://nodedesk.zendesk.com/hc/en-us/articles/202381391' );
 	$help_link = sprintf( __( '<a href="%1$s">Update plugins</a>. <a href="%2$s">More information</a>.' ), self_admin_url( 'update-core.php' ), $help_url );
 
 	$text = sprintf( esc_html__( 'Plugin "%1$s" has been deactivated as it requires a current Aihrus Framework. Once corrected, "%1$s" can be activated. %2$s' ), WPSP_NAME, $help_link );